Move `StructureClass` to it's own header file
